Installation helper script
This loads and configures the source via Github so that it can be easily updated. You'll still have to make a configuration file and adjust the boot settings but hey, all the hard work will have been done for you. It isn't ready yet so come back in a few days.